+++ tests/src/test/scala/de/knutwalker/akka/typed/TypedSpec.scala
@@ -16,20 +16,20 @@
 
 package de.knutwalker.akka.typed
 
-import akka.actor.{ Actor, Deploy, Terminated, PoisonPill, ActorSystem, Inbox }
+import akka.actor.{ Actor, Deploy, Terminated, PoisonPill, ActorSystem }
+import akka.pattern.AskTimeoutException
 import akka.routing.NoRouter
 import akka.util.Timeout
 import org.specs2.concurrent.ExecutionEnv
 import org.specs2.execute._
 import org.specs2.execute.Typecheck._
-import org.specs2.matcher.Matcher
 import org.specs2.matcher.TypecheckMatchers._
 import org.specs2.mutable.Specification
 import org.specs2.specification.AfterAll
 
-import scala.annotation.tailrec
-import scala.concurrent.{ Await, TimeoutException }
 import scala.concurrent.duration._
+
+import java.util.regex.Pattern
 
 object TypedSpec extends Specification with AfterAll {
   sequential
@@ -42,16 +42,8 @@ object TypedSpec extends Specification with AfterAll {
   case class SomeOtherMessage(msg: String)
 
   implicit val system = ActorSystem("test")
-  val inbox    = createInbox(system)
+  val inbox    = CreateInbox()
   val inboxRef = inbox.getRef()
-
-  // https://github.com/akka/akka/issues/15409
-  @tailrec
-  def createInbox(sys: ActorSystem): Inbox = {
-    try Inbox.create(system) catch {
-      case cee: ClassCastException ⇒ createInbox(sys)
-    }
-  }
 
   case class MyActor(name: String) extends TypedActor.Of[TestMessage] {
     def typedReceive: TypedReceive = {
@@ -94,31 +86,35 @@ object TypedSpec extends Specification with AfterAll {
   }
 
   "ask support of typed actors" should {
-    implicit val timeout: Timeout = 100.millis
 
     "ask the question and the return a future" >> { implicit ee: ExecutionEnv ⇒
+      implicit val timeout: Timeout = (100 * ee.timeFactor).millis
       val ref = ActorOf(TypedActor[Baz](m ⇒ m.replyTo ! SomeOtherMessage(m.msg)))
       (ref ? Baz("foo")) must be_==(SomeOtherMessage("foo")).await
     }
 
     "respect the timeout" >> { implicit ee: ExecutionEnv ⇒
+      implicit val timeout: Timeout = (100 * ee.timeFactor).millis
       val ref = ActorOf(TypedActor[Baz](_ ⇒ ()), "discarder")
       val expectedMessage = TimeoutMessage(ref)
-      patchedAwait(ref)(be_===(expectedMessage))
+      val matchMessage = s"^${Pattern.quote(expectedMessage)}$$"
+      (ref ? Baz("foo")) must throwAn[AskTimeoutException](message = matchMessage).await
     }
 
     "fail with an invalid timeout" >> { implicit ee: ExecutionEnv ⇒
-      val ref = ActorOf(TypedActor[Baz](m ⇒ m.replyTo ! SomeOtherMessage(m.msg)))
       implicit val timeout: Timeout = -100.millis
+      val ref = ActorOf(TypedActor[Baz](m ⇒ m.replyTo ! SomeOtherMessage(m.msg)))
       (ref ? Baz("foo")) must throwAn[IllegalArgumentException].like {
         case e ⇒ e.getMessage must startWith("Timeout length must not be negative, question not sent")
       }.await
     }
 
     "fail when the target is already terminated" >> { implicit ee: ExecutionEnv ⇒
+      implicit val timeout: Timeout = (100 * ee.timeFactor).millis
       val ref = kill(ActorOf(TypedActor[Baz](m ⇒ m.replyTo ! SomeOtherMessage(m.msg))))
       val expectedMessage = s"Recipient[$ref] had already been terminated."
-      patchedAwait(ref)(startWith(expectedMessage))
+      val matchMessage = s"^${Pattern.quote(expectedMessage)}.*$$"
+      (ref ? Baz("foo")) must throwAn[AskTimeoutException](message = matchMessage).await
     }
   }
 
@@ -221,23 +217,6 @@ object TypedSpec extends Specification with AfterAll {
     "withDeploy()" >> {
       props.withDeploy(Deploy("foo")).untyped must be_=== (untyped.withDeploy(Deploy("foo")))
       props.withDeploy(Deploy("foo")) must be_=== (untyped.withDeploy(Deploy("foo")).typed)
-    }
-  }
-
-  def patchedAwait(ref: ActorRef[Baz])(m: ⇒ Matcher[String])(implicit timeout: Timeout): Result = {
-    // cant use .await matcher since s.c.Await.result throws a TimeoutException
-    // and akka AskTimeout is <: TimeoutException as well
-    // see https://github.com/etorreborre/specs2/pull/417
-    try {
-      scala.concurrent.Await.result(ref ? Baz("foo"), 1.second)
-      Success()
-    } catch {
-      // check if exception is a direct instance of TimeoutException (await timeout)
-      // and not of a subclass (akka timeout)
-      case te: TimeoutException if te.getClass == classOf[TimeoutException] ⇒
-        Failure(s"Timeout after 1 second")
-      case other: Throwable ⇒
-        createExpectable(other.getMessage).applyMatcher(m).toResult
     }
   }